Consistent implementation of (inverse) Fourier transforms.
Asserts a uniform implementation of (inverse) Fourier transforms. Using the new methods over the ones provided by scipy or numpy directly makes sure that the internal implementation is consistent. It also improves maintainability as changing the internal implementation of the (inverse) Fourier transform only needs to be done in one file (`fourier.py`).